add denominations

+++ PrizeBondStore/Denomination.cs
@@ -21,7 +21,7 @@ namespace PrizeBondStore
         }
     
         public int DenominationId { get; set; }
-        public decimal Type { get; set; }
+        public int Type { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
         public virtual ICollection<PrizeBond> PrizeBonds { get; set; }
